==> Manning/Phishing_Websites_Detection/notebooks/1. Loading and understanding the Phishing Websites dataset.ipynb <==
# Filter the uneccesary warnings
import warnings
warnings.filterwarnings("ignore")

# Import pandas and numpy
import pandas as pd
import numpy as np

# Fix the random seed
np.random.seed(7)

Note that you can supply any integer number to `np.random.seed`. 

# Load the dataset
data = pd.read_csv("Phishing.csv")
data.head(10)

While the `head` function returns the first N (the number you supply to it) rows from the DataFrame, the `sample` function randomly samples N (the number you supply to it) rows from the DataFrame. I often find it useful to use the `sample` function. 

data.sample(10)

Note that the above table is large enough to _not_ fit in the regular displays. You may meet with a dataset that has even more number of columns than the current dataset has. So, in those situations it's good to make use of the `T` attribute. 

data.sample(10).T

`pandas` makes it super easy to investigate the primary facts about your DataFrame. You can quickly count the number of rows and columns of the DataFrame like so - 

# Data dimension
data.shape

The names of the columns in the DataFrame are descriptive. You may also encounter datasets where the column names are anonymized for the sake of user privacy. You can get the names of the columns as easily as - 

# Data columns
data.columns

The labels of the dataset are present in the `Result` column. Let's quickly find out how many unique labels/classes are there in the dataset - 

data["Result"].unique()

where, -1 denotes "Phishing" and 1 denotes "Normal". As there are only classes present here, it can be termed as a binary classification problem (classification since we want to classify if a website is phishing or not given its features).
