==> Manning/Phishing_Websites_Detection/notebooks/3. Cleaning the class labels and inspecting for missing values.ipynb <==
# Filter the uneccesary warnings
import warnings
warnings.filterwarnings("ignore")

# Import pandas and numpy
import pandas as pd
import numpy as np

# Fix the random seed
np.random.seed(7)

# Load the dataset
data = pd.read_csv("../../data/Phishing.csv")
data.head(10).T

It is not a good practice to build machine learning models where the labels are encoded as negative values. It affects the performance of the models. So, you would want to change the -1 values to 0. 

# Rebase the values
data.rename(columns={"Result": "Class"}, inplace=True)

data["Class"] = data["Class"].map({-1:0, 1:1})
data["Class"].unique()

We can look for missing values in the dataset like so -

data.isna().sum()

As I had mentioned that missing values can be present in a lot of various forms. Refer to the article I mentioned in the previous task's solution to know more about this. 

Let's now split the dataset in a 80:20 ratio. 

from sklearn.model_selection import train_test_split

X = data.iloc[:,0:30].values.astype(int)
y = data.iloc[:,30].values.astype(int)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=np.random.seed(7))

Let's serialize the splits as well. Remember that our splits are now nothing but `numpy` arrays and those can be easily serialized like the following. 

# Serialize the numpy arrays
np.save("X_train.npy", X_train), np.save("y_train.npy", y_train)
np.save("X_test.npy", X_train), np.save("y_test.npy", y_train)

You can safely ignore the output. 

X_train.shape
