==> Manning/Phishing_Websites_Detection/notebooks/5. Training a neural network and making it perform better.ipynb <==
In this final solution, we will be incorporating neural networks to see if we can boost the modeling performance. As the previous solution, we will first start by importing the necessary modules and will then load the dataset. 

# Filter the uneccesary warnings
import warnings
warnings.filterwarnings("ignore")

# Import numpy
import numpy as np

# Fix the random seed
np.random.seed(7)

# Load the numpy arrays which will be our datasets from now
X_train, y_train = np.load("X_train.npy", allow_pickle=True), np.load("y_train.npy", allow_pickle=True)
X_test, y_test = np.load("X_test.npy", allow_pickle=True), np.load("y_test.npy", allow_pickle=True)

# Other imports
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.optimizers import *
from wandb.keras import WandbCallback
import wandb

# Fix TensorFlow's random see for better reproducibility
import tensorflow as tf
tf.random.set_seed(666)

I find it convenient to wrap my model definition inside a function and return a compiled version of the model from the function. 

# Model building using the Sequential API
def get_training_model(data):
    model = Sequential()

    model.add(Dense(40, activation="relu",
              kernel_initializer="uniform",input_dim=data.shape[1]))
    model.add(Dense(30, activation="relu",
              kernel_initializer="uniform"))
    model.add(Dense(1, activation="sigmoid"))

    model.compile(loss="binary_crossentropy", optimizer=Adam(), metrics=["accuracy"])
    return model

When using the `ReLU` activation function it's good to use the `He_Uniform` scheme to initialize the weights of your neural network and that can be set using the `kernel_initializer` argument. 

# Instantiate the model and print its summary
model = get_training_model(X_train)
model.summary()

Early Stopping is a good way to prevent overfitting and in many cases you won't need to train the model for all the epochs. If you see that you model's performance is not as expected the training will get stopped after fixed interval (which will be set by you) and the best model within those many epochs will be returned. Let's define an `EarlyStopping` callback. 

from tensorflow.keras.callbacks import EarlyStopping

es_cb = EarlyStopping(monitor="loss", 
    patience=5, # number of epochs to consider
    restore_best_weights=True, # restore the best weights when loss stops enhancing
    verbose=1)

Let's start training our shallow network. 

from sklearn.metrics import accuracy_score, precision_recall_fscore_support, classification_report
import time 

def train_network(model, name):
    # Initialize Weights and Biases
    wandb.init(project="phishing-websites-detection", name=name)
    
    start = time.time()
    history = model.fit(X_train, y_train, batch_size=64, epochs=128, verbose=1, \
        callbacks=[es_cb, WandbCallback()])
    end = time.time()-start
    prediction = model.predict_classes(X_test)
    wandb.log({"accuracy":accuracy_score(y_test, prediction)*100.0,\
               "precision": precision_recall_fscore_support(y_test, prediction, average="macro")[0],
               "recall": precision_recall_fscore_support(y_test, prediction, average="macro")[1],
               "training_time":end})
    print("Accuracy score of the Logistic Regression classifier with default hyperparameter values {0:.2f}%"\
                  .format(accuracy_score(y_test, prediction)*100.))
    print("\n")
    print("----Classification report of the Logistic Regression classifier with default hyperparameter value----")
    print("\n")
    print(classification_report(y_test, prediction, target_names=["Phishing Websites", "Normal Websites"]))

I would encourage you to specify other things like the data, number of epochs, batch size as arguments to the function to have more flexibility and control. 

train_network(model, "neural-network")

We already have a training accuracy of 97%. Can we push this even further? 

**TDLHBA** is technique introduced [in this paper](https://dl.acm.org/citation.cfm?id=3227655). We will use hyperparameter values as presented in the paper to see the performance enhancement of the model.

# Building the model with the same topology as specified in the above-mentioned paper

model_TDLHBA = Sequential()

model_TDLHBA.add(Dense(40, activation="relu",
          kernel_initializer="uniform", input_dim=X_train.shape[1]))
model_TDLHBA.add(Dense(30, activation="relu",
          kernel_initializer="uniform"))
model_TDLHBA.add(Dense(1, activation="sigmoid"))

adam = Adam(lr=0.0017470)
model_TDLHBA.compile(loss="binary_crossentropy", optimizer=adam, metrics=["accuracy"])

wandb.init(project="phishing-websites-detection", name="neural-network-tdlhba")

start = time.time()
history_TDLHBA = model_TDLHBA.fit(X_train, y_train, batch_size=10, epochs=100, 
                                  verbose=1, callbacks=[es_cb, WandbCallback()])

end = time.time() - start
prediction = model_TDLHBA.predict_classes(X_test)
wandb.log({"accuracy":accuracy_score(y_test, prediction)*100.0,\
           "precision": precision_recall_fscore_support(y_test, prediction, average="macro")[0],
           "recall": precision_recall_fscore_support(y_test, prediction, average="macro")[1],
           "training_time":end})
print("Accuracy score of the Logistic Regression classifier with default hyperparameter values {0:.2f}%"\
                  .format(accuracy_score(y_test, prediction)*100.))
print("\n")
print("----Classification report of the Logistic Regression classifier with default hyperparameter value----")
print("\n")
print(classification_report(y_test, prediction, target_names=["Phishing Websites", "Normal Websites"]))

The performance is by far the best one.
